requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.3
cached-property==2.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
et_xmlfile==2.0.0
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
lxml==6.0.2
multidict==6.7.0
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
propcache==0.4.1
pyee==13.0.0
python-dateutil==2.9.0.post0
python-Wappalyzer @ file:///home/senju/Projects/dinosaur-engine/python-Wappalyzer
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
